Move bootloader stuff from Blivet to InstallerStorage.
This moves a number of things out of Blivet class and into
InstallerStorage, particularly:

    _bootloader (attr)
    bootloader (prop)
    update_bootloader_disk_list
    set_up_bootloader
    boot_disk (prop)
    boot_device (prop)
    bootloader_device (prop)
    boot_fstypes (prop)
    default_boot_fstype (prop)
    set_default_boot_fstype

Blivet.get_fstype had to also be reduced a bit for the specific
default_boot_fstype call which was moved into InstallerStorage.
